+++ pytket/docs/changelog.rst
@@ -1,7 +1,18 @@
 Changelog
 =========
 
-1.4.0 (July 2022)
+x.y.z (unreleased)
+------------------
+
+Minor new features:
+
+* Add support for TK2 gate in ``KAKDecomposition``.
+* ``Transform.ThreeQubitSquash()`` can now use TK2 gates as an alternative to CX
+  gates.
+* ``Unitary3qBox.get_circuit()`` decomposes the circuit using (at most 15) TK2
+  gates.
+
+1.4.1 (July 2022)
 -----------------
 
 Minor new features: